Refactor extensions service and return all filtered results
